# backend/docs/2026-07-20-openapi-draft.yaml
openapi: 3.1.0
info:
  title: Agency OS API
  version: 0.1.0
  description: Draft API specification for the first backend vertical slice.
servers:
  - url: http://localhost:4000/api/v1
security:
  - bearerAuth: []
paths:
...
  /content-assets/{id}:
    get:
      summary: Get content asset with workflow execution data
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Content asset found
    patch:
      summary: Update domain content brief
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Content asset updated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...